+++ components/BlogFooter.vue
@@ -4,7 +4,10 @@
     <h2>Comments</h2>
     <p>
       If you have any questions or comments about this post, please feel free to
-      leave a comment below.
+      leave a comment <span v-if="frontmatter.comments === 'giscus'">below</span
+      ><span v-if="frontmatter.comments === 'activitypub'"
+        >on <a :href="frontmatter.posts">this post</a></span
+      >.
     </p>
 
     <noscript
@@ -14,7 +17,10 @@
         JavaScript.</small
       ></noscript
     >
-    <div id="giscusContainer"></div>
+    <MastodonComments
+      v-if="frontmatter.comments === 'activitypub'"
+    ></MastodonComments>
+    <Giscus v-if="frontmatter.comments === 'giscus'"></Giscus>
     <br />
     <hr />
     <p class="dialog-buttons" v-if="frontmatter.prev || frontmatter.next">
@@ -40,31 +46,10 @@
 
 <script setup lang="ts">
 import { useData } from "vitepress";
-import { onMounted } from "vue";
+import Giscus from "./GiscusComments.vue";
+import MastodonComments from "./MastodonComments.vue";
 
-const { theme, frontmatter } = useData();
-
-onMounted(() => {
-  // Load GitHub Comments script dynamically
-  const script = document.createElement("script");
-  script.src = "https://giscus.app/client.js";
-  script.async = true;
-  script.setAttribute("data-repo", theme.value.blog.data.repo);
-  script.setAttribute("data-repo-id", theme.value.blog.data.repoid);
-  script.setAttribute("data-category", "Blog Comments");
-  script.setAttribute("data-category-id", theme.value.blog.data.categoryid);
-  script.setAttribute("data-mapping", "title");
-  script.setAttribute("data-strict", "0");
-  script.setAttribute("data-reactions-enabled", "1");
-  script.setAttribute("data-emit-metadata", "0");
-  script.setAttribute("data-input-position", "bottom");
-  script.setAttribute("data-theme", "preferred_color_scheme");
-  script.setAttribute("data-lang", "en");
-  script.setAttribute("crossorigin", "anonymous");
-
-  // Append the script to the document
-  document.getElementById("giscusContainer")?.appendChild(script);
-});
+const { frontmatter } = useData();
 </script>
 
 <style lang="scss">